When we get an abiguity problem with a multi-statement closure return type, it is almost always the case that the user didn't know what the rules are between single expression and multistatement closures, and they often don't know how to fix the problem.
Address this by doing some heroics when we detect this situation.  We now go dive
into the closure body, type check the explicit returns within it, and can usually
divine the right answer.  When we do that, generate a fixit hint that generates a
modification to the existing signature, or synthesizes the entire signature from
scratch.  This addresses:
<rdar://problem/22123191> QoI: multi-line closure with failure to infer result type should add a fixit