Disable PlaygroundSupport when building with Xcode.
Right now, build-script-impl always tries to build all
PlaygroundSupport targets even if some SDKs are disabled. (That isn't
going to work, of course.) The right fix is to figure out which
targets /should/ be built given the SDKs that are enabled, but today's
fix is just to remove it from the preset that uses Xcode (which can
only build for macOS).

(The irony of disabling a subproject that only builds with Xcode when
we use Xcode to build another subproject is not lost on me.)