Always resolve overrides when doing member lookups.
Otherwise, we'll treat lookups on generic parameters as ambiguous,
since we don't have a concrete nominal type to use (coincidentally)
as the "base" of each lookup result.

It's possible we can get rid of this condition altogether, but I don't
want to rock the boat for unqualified lookup right now.

rdar://problem/23087787