[update-checkout] Accept scheme along side branch to set the scheme argument.
In order to not break the bots, I kept on accepting branch to specify the branch
scheme that defines the branches used for each repo. This now allows us to
accept both arguments to specify the branch scheme. Once the bots have migrated,
I will remove support for specifying the branch scheme via --branch.