SILGen: Don't pretend to handle ObjC APIs that "lie" about block return optionality.
If an ObjC API claims to return a nonnull object of a bridged type, such
as blocks, then we're already screwed since we don't take that
possibility into account when bridging to a Swift function. Attempting
to peephole that case only generates broken code attempting to bitcast
() -> () to Optional<() -> ()>, which is invalid due to the abstraction
change between () -> () and Optional<T>. Fixes SR-2331.