Fix UnsafeRawPointer.load to use Builtin.loadRaw.
This API should have been converted to use the
new Builtin before it was introduced here:
commit a41484ea2bf090f4018d32695abbfd6be9a95acd
Author: Andrew Trick <atrick@apple.com>
Date:   Fri Jul 22 13:32:08 2016
    Add UnsafeRawPointer type and API. (#3677)

But it looks like that was dropped during during some local merge. It's not
likely to be a problem until developers start adopting the new API and the
optimizer starts using TBAA aggressively. The other "raw" API's are binding
memory or using memmove, so those are safe.

rdar:23406272.