Uninhabited != NoReturn
Previously, isNever would return true for an equivalence class of
uninhabited enums.  The rest of the compiler, however, is using this in
places that actually expect just the Never type.  This means that code
like this would compile properly

enum Uninhabited {}
func do() -> Uninhabited { /* No body here */ }

and we wouldn’t diagnose the missing return.