[docs] LibraryEvolution: 'final' can be added or removed at will.
Slava came up with a way to do this: provide a single point of entry
that turns around and performs vtable dispatch if necessary. This
makes 'final' into an internal guarantee (much like a "performance
assertion") rather than part of a module's interface.

We can reconsider this, of course, but take it as plan of record for
now.