AST: Introduce new ArchetypeMapping class
Right now after building a GenericSignature, we go ahead and store a
bunch of state back into the original GenericParamList, which stinks.

Instead, let's build an ArchetypeMapping instance, storing the
mapping between interface types and archetypes in this generic
context.

This will replace the GenericTypeParamDecl::getArchetype() method,
and also eventually allow us to stop serializing GenericParamLists,
in the same way that we never serialize TypeReprs.

For now, we just build the ArchetypeMapping and perform some checks
in the ASTVerifier, but otherwise do nothing else with it.