[Parser] Correct the fixit for protocol composition's syntax change (#4483)
* [Parser] Correct a fixit to preserve the trailing content after replacing Protocol<A, B>. rdar://27992964

When migrating, we found our fixit to replace the old protocol composition syntax, namely "Protocol<A, B>",
to the new syntax, "A & B", does not preserve the trailing content after '>'. For instance, we replace "Protocol<A, B>?"
with "A & B". This patch fixes the issue by inserting whatever after '>' in the old syntax to the new
replacement string. I consider this as a hack; the root-cause fix should be in the lexer to smartly separate
'>' and '?' as two tokens instead of one.

* [test] Update existing test.